requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==5.1.4
djangorestframework==3.15.2
gitdb==4.0.11
GitPython==3.1.43
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
pbr==6.1.0
Pygments==2.18.0
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sqlparse==0.5.3
stevedore==5.4.0
urllib3==2.2.3
psycopg2-binary
gunicorn